MARS-76 Create `AccountMenu` component
