+++ client/src/components/AccountMenu/index.tsx
@@ -1,0 +1,111 @@
+import React from "react";
+
+// Existing and custom components
+import {
+  Avatar,
+  Flex,
+  Menu,
+  MenuButton,
+  MenuDivider,
+  MenuGroup,
+  MenuItem,
+  MenuList,
+  Tag,
+  Text,
+} from "@chakra-ui/react";
+import Icon from "@components/Icon";
+
+// Routing and navigation
+import { useNavigate } from "react-router-dom";
+
+// Utility functions and libraries
+import { getData } from "@database/functions";
+import { useToken } from "src/authentication/useToken";
+import dayjs from "dayjs";
+import FileSaver from "file-saver";
+import slugify from "slugify";
+
+export const AccountMenu = () => {
+  const navigate = useNavigate();
+
+  const [token, setToken] = useToken();
+
+  const performBackup = () => {
+    // Retrieve all data stored in the system
+    getData(`/system/backup`).then((response) => {
+      FileSaver.saveAs(
+        new Blob([JSON.stringify(response, null, "  ")]),
+        slugify(`backup_${dayjs(Date.now()).toJSON()}.json`)
+      );
+    });
+  };
+
+  const performLogout = () => {
+    // Invalidate the token and refresh the page
+    setToken({
+      name: token.name,
+      orcid: token.orcid,
+      id_token: "",
+    });
+    navigate(0);
+  };
+
+  return (
+    <Flex gap={"4"} align={"center"} h={"100%"}>
+      <Menu>
+        <MenuButton h={"100%"} _hover={{ bg: "gray.200" }}>
+          <Flex
+            direction={"row"}
+            align={"center"}
+            gap={"2"}
+            p={"1"}
+            ml={"2"}
+            mr={"2"}
+          >
+            <Avatar name={token.name} size={"sm"} />
+            <Text size={"xs"} fontWeight={"semibold"}>
+              {token.name}
+            </Text>
+            <Icon name={"c_down"} />
+          </Flex>
+        </MenuButton>
+
+        {/* List of drop-down menu items */}
+        <MenuList ml={"2"} mr={"2"}>
+          <MenuGroup>
+            <Flex p={"4"} pt={"2"} pb={"2"} gap={"4"} direction={"column"}>
+              <Text fontWeight={"semibold"}>Hello {token.name.split(" ")[0]}!</Text>
+              <Tag colorScheme={"green"}>{token.orcid}</Tag>
+            </Flex>
+          </MenuGroup>
+          <MenuDivider />
+
+          <MenuGroup title={"System"}>
+            <MenuItem onClick={() => performBackup()}>
+              <Flex direction={"row"} align={"center"} gap={"4"} ml={"2"}>
+                <Icon name={"download"} />
+                Backup
+              </Flex>
+            </MenuItem>
+            <MenuItem onClick={() => navigate(`/settings`)}>
+              <Flex direction={"row"} align={"center"} gap={"4"} ml={"2"}>
+                <Icon name={"settings"} />
+                Settings
+              </Flex>
+            </MenuItem>
+          </MenuGroup>
+          <MenuDivider />
+
+          <MenuGroup title={"Account"}>
+            <MenuItem onClick={() => performLogout()}>
+              <Flex direction={"row"} align={"center"} gap={"4"} ml={"2"}>
+                <Icon name={"exit"} />
+                Logout
+              </Flex>
+            </MenuItem>
+          </MenuGroup>
+        </MenuList>
+      </Menu>
+    </Flex>
+  );
+};
